requirements.txt
autopep8==1.4.3
certifi==2020.12.5
chardet==4.0.0
Click==7.0
dominate==2.3.5
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Script==2.0.6
gunicorn==19.9.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
newsapi-python==0.2.6
numpy==1.20.2
Pillow==8.2.0
pycodestyle==2.5.0
python-resize-image==1.1.19
requests==2.25.1
scipy==1.6.3
urllib3==1.26.4
visitor==0.1.3
Werkzeug==1.0.1
